Add endpoint to Federator API, cleanup (#1189)
* Add endpoint to Federator API, cleanup

* rename API to Api

* remove search endpoint

It being here doesn't give us a lot, still needs some answers and
tweaks. We will need a while to get there anyways.
