+++ services/federator/package.yaml
@@ -12,14 +12,12 @@ dependencies:
 - aeson
 - base
 - bilge
-- brig-types
 - bytestring-conversion
 - data-default
 - email-validate
 - errors
 - exceptions
 - extended
-- galley-types
 - imports
 - lens
 - metrics-core
@@ -38,6 +36,8 @@ dependencies:
 - wai
 - wai-utilities
 - warp
+- wire-api
+- wire-api-federation
 library:
   source-dirs: src
 executables:
