+++ services/federator/src/Federator/API.hs
@@ -17,53 +17,42 @@
 -- You should have received a copy of the GNU Affero General Public License along
 -- with this program. If not, see <https://www.gnu.org/licenses/>.
 
-module Federator.API where
+module Federator.API
+  ( Api (..),
+    module Fed,
+  )
+where
 
-import Brig.Types.Client.Prekey
-import Brig.Types.Test.Arbitrary ()
-import Data.Aeson.TH (deriveJSON)
-import Data.Handle (Handle (..))
-import Data.Id (UserId)
-import Data.Qualified
-import Federator.Util
+import Data.Id (ConvId, UserId)
+import Data.Qualified (Qualified)
 import Imports
 import Servant.API
 import Servant.API.Generic
-import Test.QuickCheck
+import Wire.API.Federation.API.Conversation as Fed hiding (Api)
+import Wire.API.Federation.Event as Fed
+import Wire.API.User.Client.Prekey (PrekeyBundle)
 
-data API route = API
-  { _gapiSearch ::
-      route
-        :- "i"
-        :> "search"
-        -- QUESTION: what exactly should the query be? text + domain?
-        :> QueryParam' [Required, Strict] "q" (Qualified Handle)
-        :> Get '[JSON] FUser,
-    _gapiPrekeys ::
+data Api route = Api
+  { _gapiPrekeys ::
       route
         :- "i"
         :> "users"
-        :> Capture "fqu" (Qualified UserId)
+        :> Capture "id" (Qualified UserId)
         :> "prekeys"
-        :> Get '[JSON] PrekeyBundle
+        -- FUTUREWORK(federation):
+        -- this should return a version of PrekeyBundle with qualified UserId,
+        -- defined in wire-api-federation
+        :> Get '[JSON] PrekeyBundle,
+    _gapiJoinConversationById ::
+      route
+        :- "i"
+        :> "conversations"
+        :> Capture "cnv" (Qualified ConvId)
+        :> "join"
+        :> ReqBody '[JSON] Fed.JoinConversationByIdRequest
+        :> Post '[JSON] (Fed.ConversationUpdateResult Fed.MemberJoin)
   }
   deriving (Generic)
 
--- curl http://localhost:8097/i/search?q=wef@a.com; curl http://localhost:8097/i/users/`uuid`@example.com/prekeys
-
 ----------------------------------------------------------------------
--- TODO: add roundtrip tests for *HttpApiData, *JSON, ...
---
--- TODO: the client ids in the 'PrekeyBundle' aren't really needed here.  do we want to make a
--- new type for that, then?
-
-data FUser = FUser
-  { _fuGlobalHandle :: !(Qualified Handle),
-    _fuFQU :: !(Qualified UserId)
-  }
-  deriving (Eq, Show, Generic)
-
-deriveJSON (wireJsonOptions "_fu") ''FUser
-
-instance Arbitrary FUser where
-  arbitrary = FUser <$> arbitrary <*> arbitrary
+-- FUTUREWORK: add roundtrip tests for *HttpApiData, *JSON, ...
