+++ services/federator/src/Federator/Impl.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE RecordWildCards #-}
+
 -- This file is part of the Wire Server implementation.
 --
 -- Copyright (C) 2020 Wire Swiss GmbH <opensource@wire.com>
@@ -15,25 +18,20 @@
 -- You should have received a copy of the GNU Affero General Public License along
 -- with this program. If not, see <https://www.gnu.org/licenses/>.
 
-module Federator.Util
-  ( wireJsonOptions,
+module Federator.Impl
+  ( app,
   )
 where
 
-import Data.Aeson as Aeson
-import Imports
+import Data.Proxy
+import qualified Federator.API as API
+import Federator.Types
+import Network.Wai
+import Servant.API.Generic
+import Servant.Mock
+import Servant.Server
 
-dropPrefix :: String -> String -> Maybe String
-dropPrefix pfx str =
-  if length pfx > length str
-    then Nothing
-    else case splitAt (length pfx) str of
-      (pfx', sfx) ->
-        if pfx' /= pfx
-          then Nothing
-          else Just sfx
-
--- | This is a partial function; totality of all calls must be verified by roundtrip tests on
--- the aeson instances involved.
-wireJsonOptions :: String -> Options
-wireJsonOptions pfx = defaultOptions {fieldLabelModifier = fromJust . dropPrefix pfx . fmap toLower}
+app :: Env -> Application
+app _ = serve api (mock api Proxy)
+  where
+    api = Proxy @(ToServantApi API.Api)
